+++ src/components/BrandActionMenuButton.vue
@@ -3,8 +3,11 @@
     <v-icon>mdi-dots-vertical</v-icon>
     <v-menu activator="parent" scroll-strategy="close" transition="slide-y-transition">
       <v-list>
-        <ShareLink :overrideUrl="getShareLinkUrl" display="list-item" />
+        <v-list-subheader class="text-uppercase" :slim="true" disabled>
+          {{ $t('Common.Brand') }}
+        </v-list-subheader>
         <v-divider />
+        <ShareLink :overrideUrl="getShareLinkUrl" display="list-item" />
         <OpenFoodFactsLink facet="brand" :value="brand" display="list-item" />
       </v-list>
     </v-menu>