+++ src/components/PriceActionMenuButton.vue
@@ -9,19 +9,26 @@
             {{ $t('Common.Product') }}
           </v-list-subheader>
           <v-divider />
+          <v-list-item v-if="price.product || price.category_tag" :slim="true" prepend-icon="mdi-eye-outline" :to="getProductOrCategoryDetailUrl">
+            {{ $t('Common.Details') }}
+          </v-list-item>
           <OpenFoodFactsLink v-if="price.product" :source="price.product.source" facet="product" :value="price.product.code" display="list-item" />
           <OpenFoodFactsLink v-else-if="price.category_tag" facet="category" :value="price.category_tag" display="list-item" />
         </v-sheet>
         <!-- Price actions -->
-        <v-sheet v-if="userIsPriceOwner">
+        <v-sheet v-if="!hidePriceActions">
           <v-list-subheader class="text-uppercase" :slim="true" disabled>
             {{ $t('Common.Price') }}
           </v-list-subheader>
           <v-divider />
-          <v-list-item :slim="true" prepend-icon="mdi-pencil" @click="openEditDialog">
+          <ShareLink v-if="showPriceShare" :overrideUrl="getShareLinkUrl" display="list-item" />
+          <v-list-item :slim="true" prepend-icon="mdi-eye-outline" :to="getPriceDetailUrl">
+            {{ $t('Common.Details') }}
+          </v-list-item>
+          <v-list-item v-if="userIsPriceOwner" :slim="true" prepend-icon="mdi-pencil" @click="openEditDialog">
             {{ $t('Common.Edit') }}
           </v-list-item>
-          <v-list-item :slim="true" prepend-icon="mdi-delete" @click="openDeleteConfirmationDialog">
+          <v-list-item v-if="userIsPriceOwner" :slim="true" prepend-icon="mdi-delete" @click="openDeleteConfirmationDialog">
             {{ $t('Common.Delete') }}
           </v-list-item>
         </v-sheet>
@@ -69,6 +76,7 @@ import { useAppStore } from '../store'
 export default {
   components: {
     OpenFoodFactsLink: defineAsyncComponent(() => import('../components/OpenFoodFactsLink.vue')),
+    ShareLink: defineAsyncComponent(() => import('../components/ShareLink.vue')),
     PriceEditDialog: defineAsyncComponent(() => import('../components/PriceEditDialog.vue')),
     PriceDeleteConfirmationDialog: defineAsyncComponent(() => import('../components/PriceDeleteConfirmationDialog.vue'))
   },
@@ -78,6 +86,10 @@ export default {
       default: null
     },
     hideProductActions: {
+      type: Boolean,
+      default: false
+    },
+    hidePriceActions: {
       type: Boolean,
       default: false
     },
@@ -99,6 +111,23 @@ export default {
     ...mapStores(useAppStore),
     username() {
       return this.appStore.user.username
+    },
+    getProductOrCategoryDetailUrl() {
+      if (this.price.product) {
+        return `/products/${this.price.product.code}`
+      } else if (this.price.category_tag) {
+        return `/categories/${this.price.category_tag}`
+      }
+      return null
+    },
+    getPriceDetailUrl() {
+      return `/prices/${this.price.id}`
+    },
+    showPriceShare() {
+      return this.$route.path === this.getPriceDetailUrl
+    },
+    getShareLinkUrl() {
+      return this.getPriceDetailUrl
     },
     userIsPriceOwner() {
       return this.username && (this.price.owner === this.username)