+++ src/components/ProductActionMenuButton.vue
@@ -3,9 +3,15 @@
     <v-icon>mdi-dots-vertical</v-icon>
     <v-menu activator="parent" scroll-strategy="close" transition="slide-y-transition">
       <v-list>
-        <PriceAddLink :productCode="product.code" display="list-item" />
-        <ShareLink :overrideUrl="'/products/' + product.code" display="list-item" />
+        <v-list-subheader class="text-uppercase" :slim="true" disabled>
+          {{ $t('Common.Product') }}
+        </v-list-subheader>
         <v-divider />
+        <PriceAddLink :productCode="product.code" display="list-item" />
+        <ShareLink :overrideUrl="getShareLinkUrl" display="list-item" />
+        <v-list-item :slim="true" prepend-icon="mdi-eye-outline" :to="getProductDetailUrl">
+          {{ $t('Common.Details') }}
+        </v-list-item>
         <OpenFoodFactsLink :source="product.source" facet="product" :value="product.code" display="list-item" />
       </v-list>
     </v-menu>
@@ -30,6 +36,14 @@ export default {
       type: String,
       default: 'position:absolute;bottom:6px;right:0;'
     }
+  },
+  computed: {
+    getProductDetailUrl() {
+      return `/products/${this.product.code}`
+    },
+    getShareLinkUrl() {
+      return this.getProductDetailUrl
+    },
   }
 }
 </script>